Sync and images added

+++ Detrack.Model/Item.cs
@@ -13,7 +13,7 @@
 
 		public string Sku { get; set; }
 		public string Desc { get; set; }
-		public int? Qty { get; set; }
+		public decimal? Qty { get; set; }
 		public int Reject { get; set; }
 		public string Reason { get; set; }
 	}
